chore(crons): Move crons grouptype into `monitors/grouptype.py`
We'll end up with circular imports as we work on this project if we leave these where they were
