+++ src/sentry/issues/grouptype.py
@@ -607,36 +607,6 @@ class ProfileFunctionRegressionType(GroupType):
 
 
 @dataclass(frozen=True)
-class MonitorIncidentType(GroupType):
-    type_id = 4001
-    slug = "monitor_check_in_failure"
-    description = "Crons Monitor Failed"
-    category = GroupCategory.CRON.value
-    category_v2 = GroupCategory.OUTAGE.value
-    released = True
-    creation_quota = Quota(3600, 60, 60_000)  # 60,000 per hour, sliding window of 60 seconds
-    default_priority = PriorityLevel.HIGH
-    notification_config = NotificationConfig(context=[])
-
-
-# XXX(epurkhiser): We renamed this group type but we keep the alias since we
-# store group type in pickles
-MonitorCheckInFailure = MonitorIncidentType
-
-
-@dataclass(frozen=True)
-class MonitorCheckInTimeout(MonitorIncidentType):
-    # This is deprecated, only kept around for it's type_id
-    type_id = 4002
-
-
-@dataclass(frozen=True)
-class MonitorCheckInMissed(MonitorIncidentType):
-    # This is deprecated, only kept around for it's type_id
-    type_id = 4003
-
-
-@dataclass(frozen=True)
 class ReplayRageClickType(ReplayGroupTypeDefaults, GroupType):
     type_id = 5002
     slug = "replay_click_rage"
